requirements.txt
flask==2.3.2
flask-cors
psycopg2
waitress
python-dotenv
Werkzeug==2.3.6
